+++ app/src/user/controller/index.ts
@@ -1,0 +1,1 @@
+export * from './user.controller';
